Add guidelines for descriptive text
This based in large part on RA's style guide for incentive descriptions. It's worked well for us, but I'm open to the idea that NODE's guidelines for what the text should include can be different. (I do think the micromanagement of formatting is good, though: being inconsistent on things like capitalization of "kWh" looks bad.)

Fixes #10